requirements.txt
# Core dependencies - Updated for compatibility
langchain==0.2.16
langchain-openai==0.1.25
openai==1.51.0
pydantic==2.9.2
python-dotenv==1.0.1

# Streamlit frontend
streamlit==1.38.0
plotly==5.24.1

# Data processing
pandas==2.2.3
jsonlines==4.0.0

# Development and testing
pytest==8.3.3
black==24.8.0
flake8==7.1.1

# Optional: Enhanced logging
loguru==0.7.2